Revert "Update loaded state for images with tests"
This reverts commit c58c6191b8130f4a9635858294fd74a78ccbc692.
